Rename all .h extensions to .hpp .
