+++ solvers/Electrostatic2D3V/EquationSystems/PoissonPIC.cpp
@@ -1,4 +1,4 @@
-#include "PoissonPIC.h"
+#include "PoissonPIC.hpp"
 
 using namespace std;
