+++ solvers/SimpleSOL/EquationSystems/SOLSystem.hpp
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 //
-// File SOLSystem.h
+// File SOLSystem.hpp
 //
 // For more information, please see: http://www.nektar.info
 //
