+++ solvers/SimpleSOL/EquationSystems/SOLWithParticlesSystem.cpp
@@ -34,7 +34,7 @@
 
 #include <boost/algorithm/string/predicate.hpp>
 
-#include "SOLWithParticlesSystem.h"
+#include "SOLWithParticlesSystem.hpp"
 
 namespace Nektar {
 string SOLWithParticlesSystem::className =
