+++ solvers/SimpleSOL/Forcing/SourceTerms.hpp
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 //
-// File: SourceTerms.h
+// File: SourceTerms.hpp
 //
 // For more information, please see: http://www.nektar.info
 //
