+++ solvers/SimpleSOL/SimpleSOL.hpp
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 //
-// File: SimpleSOL.h
+// File: SimpleSOL.hpp
 //
 //
 // Description: Header for the SimpleSOL solver.
