+++ test/integration/solvers/SimpleSOL/test_SimpleSOL.hpp
@@ -4,19 +4,19 @@
 #include <boost/algorithm/string.hpp>
 #include <boost/program_options.hpp>
 #include <fstream>
-#include <vector>
 #include <gtest/gtest.h>
+#include <vector>
 
 #include <FieldUtils/Module.h>
 #include <LibUtilities/BasicConst/NektarUnivTypeDefs.hpp>
 #include <LibUtilities/BasicUtils/SharedArray.hpp>
 #include <LibUtilities/Communication/CommSerial.h>
 
-#include "SimpleSOL.h"
-#include "solver_test_utils.h"
-#include "solvers/solver_runner.hpp"
+#include "EquationSystems/SOLWithParticlesSystem.hpp"
+#include "SimpleSOL.hpp"
+#include "solver_test_utils.hpp"
 #include "solvers/solver_callback_handler.hpp"
-#include "EquationSystems/SOLWithParticlesSystem.h"
+#include "solvers/solver_runner.hpp"
 
 namespace LU = Nektar::LibUtilities;
 namespace FU = Nektar::FieldUtils;
@@ -186,14 +186,15 @@ protected:
   }
 };
 
-
-struct SOLWithParticlesMassConservationPre : public NESO::SolverCallback<SOLWithParticlesSystem> {
+struct SOLWithParticlesMassConservationPre
+    : public NESO::SolverCallback<SOLWithParticlesSystem> {
   void call(SOLWithParticlesSystem *state) {
     state->m_diag_mass_recording->compute_initial_fluid_mass();
   }
 };
 
-struct SOLWithParticlesMassConservationPost : public NESO::SolverCallback<SOLWithParticlesSystem> {
+struct SOLWithParticlesMassConservationPost
+    : public NESO::SolverCallback<SOLWithParticlesSystem> {
   std::vector<double> mass_error;
   void call(SOLWithParticlesSystem *state) {
     auto md = state->m_diag_mass_recording;
@@ -202,9 +203,9 @@ struct SOLWithParticlesMassConservationPost : public NESO::SolverCallback<SOLWit
     const double mass_total = mass_particles + mass_fluid;
     const double mass_added = md->compute_total_added_mass();
     const double correct_total = mass_added + md->get_initial_mass();
-    this->mass_error.push_back(std::fabs(correct_total - mass_total)/std::fabs(correct_total));
+    this->mass_error.push_back(std::fabs(correct_total - mass_total) /
+                               std::fabs(correct_total));
   }
 };
-
 
 #endif // SIMPLESOL_TESTS_COMMON
